requirements.txt
anyio
asgiref
certifi
charset-normalizer
crispy-bootstrap5
dj-database-url
Django
django-cors-headers
django-crispy-forms
django-environ
django-filter
djangorestframework
h11
httpcore
httpx
idna
Markdown
MouseInfo
pillow
psycopg2-binary
PyAutoGUI
PyGetWindow
PyMsgBox
pyperclip
PyRect
PyScreeze
python-dotenv
pytweening
requests
setuptools
sniffio
sqlparse
sslcommerz-lib
typing_extensions
tzdata
urllib3
wheel
whitenoise